+++ src/converse.js
@@ -28,6 +28,7 @@ import "converse-push";            // XEP-0357 Push Notifications
 import "converse-register";        // XEP-0077 In-band registration
 import "converse-roomslist";       // Show currently open chat rooms
 import "converse-rosterview";
+import "converse-forward-message"; // allows to redirect messages to other users or MUCs (XEP-0297)
 import "converse-singleton";
 import "converse-uniview";
 /* END: Removable components */
@@ -59,6 +60,7 @@ const WHITELISTED_PLUGINS = [
     'converse-register',
     'converse-roomslist',
     'converse-rosterview',
+    'converse-forward-message',
     'converse-singleton',
     'converse-uniview'
 ];